# Make revlex -> bases faster, use tuples for base elements