This improves the speed of the conversion revlex -> basis.

In addition, I changed basis elements to be tuples instead of lists, it seems more natural, for instance to be used as dictionary keys or making a set. But also because using lists requires either (a) creating a list for each basis element of each matroid (slower) or (b) aliasing mutable objects (dangerous).

Before:
```
In [1]: from matroid_database import *

In [2]: %timeit sum(1 for _ in all_matroids_bases(10, 3))
807 ms ± 12.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [3]: %time sum(1 for _ in all_matroids_bases(11, 3))
CPU times: user 33.9 s, sys: 6.98 ms, total: 33.9 s
Wall time: 33.9 s
Out[3]: 298491
```

After:
```
In [1]: from matroid_database import *

In [2]: %timeit sum(1 for _ in all_matroids_bases(10, 3))
56.2 ms ± 937 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [3]: %time sum(1 for _ in all_matroids_bases(11, 3))
CPU times: user 2.17 s, sys: 5.99 ms, total: 2.17 s
Wall time: 2.18 s
Out[3]: 298491
```